move checkpointer io test to test_io
