+++ direct/checkpointer.py
@@ -17,8 +17,7 @@ from torch.nn.parallel import DistributedDataParallel
 
 from direct.environment import DIRECT_MODEL_DOWNLOAD_DIR
 from direct.types import HasStateDict, PathOrString
-from direct.utils.io import download_url, check_is_valid_url
-
+from direct.utils.io import check_is_valid_url, download_url
 
 # TODO: Rewrite Checkpointer
 # There are too many issues with typing and mypy in the checkpointer.
